requirements.txt
typer==0.12.3
pillow==10.3.0
roboflow==1.1.28
ultralytics==8.2.14
scikit-image==0.23.2
PyYAML==6.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
